requirements.txt
# ============================================================================
# REQUIREMENTS.TXT - SISTEMA BMS DEMO - MÓDULO 1
# Dependencias para configuración base + Protocolo Modbus + Modelos
# ============================================================================

# Framework Web (para futuros módulos)
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5

# Protocolos de Comunicación - MÓDULO 1 (Modbus)
pymodbus==3.4.1

# Protocolos adicionales (para futuros módulos)
bacpypes==0.18.6
paho-mqtt==1.6.1
pysnmp==4.4.12
requests==2.31.0

# Base de Datos
SQLAlchemy==2.0.20
# sqlite3 está incluido en Python

# Utilidades del Sistema
python-dotenv==1.0.0
schedule==1.2.0
colorlog==6.7.0
psutil==5.9.5

# Desarrollo y Pruebas
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Dependencias adicionales para funcionalidad completa
dataclasses-json==0.5.14
typing-extensions==4.7.1

# ============================================================================
# NOTAS DE INSTALACIÓN
# ============================================================================
# 
# Para instalar todas las dependencias:
# pip install -r requirements.txt
#
# Para instalar solo las dependencias del Módulo 1:
# pip install pymodbus==3.4.1 python-dotenv==1.0.0 SQLAlchemy==2.0.20 colorlog==6.7.0 psutil==5.9.5
#
# Para desarrollo:
# pip install pytest==7.4.2 black==23.7.0 flake8==6.0.0
#
# ============================================================================